folder structure and spliting codes into different files
